+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.0.6] - 2025-07-28
+### Added
+- Multiple roles defined in teiler backend
 
 ## [1.0.5] - 2025-06-24
 ### Removed